% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_clean_numeric.R
\name{as_clean_numeric}
\alias{as_clean_numeric}
\title{As Clean Numeric}
\usage{
as_clean_numeric(
  x,
  removal_pattern = c("\\\\$|\\\\,|\\\\\%|\\\\#"),
  show_nas = FALSE
)
}
\arguments{
\item{x}{object to be coerced}

\item{removal_pattern}{regular expression to be used in removal of special characters}

\item{show_nas}{a logical for whether or not to print a message of up to 5 characters that are converted to NAs}
}
\value{
object of class numeric
}
\description{
Read a data directory to a data frame or a list
}
\examples{
\dontrun{

# this is perfect if you have a column in a dataframe which you would like to
# convert a poorly formated character vector of numbers to numeric
df <- df \%>\%
  mutate(
    # sometimes variables in dollars have $ and/or ,s
    gdp = as_clean_numeric(gdp),
    # sometimes these have percent signs
    percent_under30 = as_clean_numeric(percent_under_30)
  )

}

}
