% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_case.R
\name{factor_case}
\alias{factor_case}
\title{Adjust Factor Case}
\usage{
factor_case(var, case_fxn = stringr::str_to_title)
}
\arguments{
\item{var}{variable of choice; a vector of class factor}

\item{case_fxn}{a function to be used to adjust the case of var; defaults to str_to_title}
}
\value{
vector of class factor
}
\description{
Adjust the case of a factor variable.
For example make its categories title case instead of lower case while keeping the levels in the same order
}
\examples{

df <- data.frame(
  gender = as.factor(c("man", "woman", "non-binary")),
  blood_type = as.factor(c("ab", "o", "a"))
)

#to make it title case
df$gender <- factor_case(df$gender)
#to make it all caps
df$blood_type <- factor_case(df$blood_type, toupper)

}
