% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_directory.R
\name{read_data_directory}
\alias{read_data_directory}
\title{Read Data Directory}
\usage{
read_data_directory(
  directory_path,
  map_fxn = purrr::map_dfr,
  read_fxn = readr::read_csv
)
}
\arguments{
\item{directory_path}{path to the data directory}

\item{map_fxn}{the map_fxn to use; defaults to map_dfr which is ideal for creating a dataframe}

\item{read_fxn}{the function used to read in individual files; defaults to read.csv}
}
\value{
data return depends on map_fxn
}
\description{
Read a data directory to a data frame or a list
}
\examples{
\dontrun{
directory_path <- "Data/PPIC"
#to read the data into a data frame
df <- read_data_directory(directory_path, purrr::map_dfr, haven::read_sav)
#to read the data into list
data_list <- read_data_directory(directory_path, purrr::map, haven::read_sav)

#with defaults you can also just read data from csvs to a dataframe
df <- read_data_directory(directory_path)
}

}
