% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_colors.R
\name{visualize_colors}
\alias{visualize_colors}
\title{Visualize Colors}
\usage{
visualize_colors(color_pal)
}
\arguments{
\item{color_pal}{a vector of character corresponding to colors to plot}
}
\value{
a barplot (graphics version, not ggplot)
}
\description{
This function plots supplied colors in a barplot so the user can get a quick sense of what they look like
}
\examples{

visualize_colors(c("dodgerblue", "darkgrey"))

\dontrun{

# for those who use viridis
visualize_colors(viridis::viridis(5))

}

}
